backend/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib
pandas
openpyxl
python-dotenv
